+++ stroom-core/src/main/java/stroom/core/receive/AttributeMapFilterFactory.java
@@ -1,0 +1,62 @@
+/*
+ * Copyright 2017 Crown Copyright
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package stroom.core.receive;
+
+import stroom.docref.DocRef;
+import stroom.receive.common.AttributeMapFilter;
+import stroom.receive.common.DataReceiptPolicyAttributeMapFilterFactory;
+import stroom.receive.common.FeedStatusAttributeMapFilter;
+
+import javax.inject.Inject;
+import javax.inject.Singleton;
+import java.util.Objects;
+import java.util.concurrent.atomic.AtomicReference;
+
+@Singleton
+public class AttributeMapFilterFactory {
+    private final ReceiveDataConfig receiveDataConfig;
+    private final DataReceiptPolicyAttributeMapFilterFactory dataReceiptPolicyAttributeMapFilterFactory;
+    private final FeedStatusAttributeMapFilter feedStatusAttributeMapFilter;
+
+    private volatile AttributeMapFilter attributeMapFilter;
+    private final AtomicReference<String> lastPolicyUuid = new AtomicReference<>();
+
+    @Inject
+    public AttributeMapFilterFactory(final ReceiveDataConfig receiveDataConfig,
+                                     final DataReceiptPolicyAttributeMapFilterFactory dataReceiptPolicyAttributeMapFilterFactory,
+                                     final FeedStatusAttributeMapFilter feedStatusAttributeMapFilter) {
+        this.receiveDataConfig = receiveDataConfig;
+        this.dataReceiptPolicyAttributeMapFilterFactory = dataReceiptPolicyAttributeMapFilterFactory;
+        this.feedStatusAttributeMapFilter = feedStatusAttributeMapFilter;
+    }
+
+    public AttributeMapFilter create() {
+        final String receiptPolicyUuid = receiveDataConfig.getReceiptPolicyUuid();
+        final String last = lastPolicyUuid.get();
+        if (attributeMapFilter == null || !Objects.equals(last, receiptPolicyUuid)) {
+            lastPolicyUuid.compareAndSet(last, receiptPolicyUuid);
+
+            if (receiptPolicyUuid != null && receiptPolicyUuid.length() > 0) {
+                attributeMapFilter = dataReceiptPolicyAttributeMapFilterFactory.create(new DocRef("RuleSet", receiptPolicyUuid));
+            } else {
+                attributeMapFilter = feedStatusAttributeMapFilter;
+            }
+        }
+
+        return attributeMapFilter;
+    }
+}