+++ stroom-core/src/main/java/stroom/core/receive/ReceiveDataModule.java
@@ -18,6 +18,7 @@ package stroom.core.receive;
 
 import com.google.inject.AbstractModule;
 import stroom.proxy.repo.FileSetProcessor;
+import stroom.receive.common.FeedStatusService;
 import stroom.util.io.BufferFactory;
 import stroom.receive.common.RequestHandler;
 
@@ -27,5 +28,6 @@ public class ReceiveDataModule extends AbstractModule {
         bind(BufferFactory.class).to(BufferFactoryImpl.class);
         bind(FileSetProcessor.class).to(DataStoreFileSetProcessor.class);
         bind(RequestHandler.class).to(ReceiveDataRequestHandler.class);
+        bind(FeedStatusService.class).to(FeedStatusServiceImpl.class);
     }
 }