+++ stroom-node/stroom-node-impl/src/main/java/stroom/node/impl/RemoteStatusServiceRPC.java
@@ -1,6 +1,8 @@
 package stroom.node.impl;
 
 import com.caucho.hessian.server.HessianServlet;
+import stroom.proxy.status.remote.GetStatusRequest;
+import stroom.proxy.status.remote.GetStatusResponse;
 
 import javax.inject.Inject;