+++ stroom-receive/stroom-receive-common/src/main/java/stroom/receive/common/DataReceiptPolicyAttributeMapFilterFactory.java
@@ -1,0 +1,7 @@
+package stroom.receive.common;
+
+import stroom.docref.DocRef;
+
+public interface DataReceiptPolicyAttributeMapFilterFactory {
+    AttributeMapFilter create(DocRef policyRef);
+}