+++ stroom-receive/stroom-receive-common/src/main/java/stroom/receive/common/PermissiveAttributeMapFilter.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 Crown Copyright
+ * Copyright 2017 Crown Copyright
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -12,12 +12,16 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
+ *
  */
 
-package stroom.core.feed;
+package stroom.receive.common;
 
-import java.io.Serializable;
+import stroom.meta.shared.AttributeMap;
 
-public class RemoteRequest implements Serializable {
-    private static final long serialVersionUID = -3023517525631202886L;
+public class PermissiveAttributeMapFilter implements AttributeMapFilter {
+    @Override
+    public boolean filter(AttributeMap attributeMap) {
+        return true;
+    }
 }