+++ stroom-receive/stroom-receive-common/src/main/java/stroom/receive/common/StroomStreamException.java
@@ -19,6 +19,7 @@ package stroom.receive.common;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import stroom.meta.shared.StandardHeaderArguments;
+import stroom.proxy.StroomStatusCode;
 
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;